requirements.txt
dj-database-url==0.3.0
Django==1.9
django-postgrespool==0.3.0
gunicorn==19.3.0
psycopg2==2.6
SQLAlchemy==1.0.4
whitenoise==2.0.2
livereload
djangorestframework==3.3.2
django-extensions==1.6.1
django-rest-swagger==0.3.4
requests==2.9.1
python-social-auth==0.2.13
django-rest-auth==0.6.0
django-cors-headers==1.1.0
django-markupfield==1.4.0
markdown==2.6.5
